+++ examples/data-sources/meraki_appliance_firewalled_service/data-source.tf
@@ -1,0 +1,4 @@
+data "meraki_appliance_firewalled_service" "example" {
+  network_id = "L_123456"
+  service    = "ICMP"
+}