+++ examples/resources/meraki_appliance_firewalled_service/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_firewalled_service.example "<network_id>,<service>"